requirements.txt
# API and Core
fastapi==0.115.0
uvicorn==0.31.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.10

# Database
pymongo==4.8.0
redis==5.0.8
hyperspacedb==3.0.3

# Background Tasks
celery==5.4.0

# Document Processing and RAG
pdf2image==1.17.0
python-docx==1.1.2
langchain-core>=0.3.15
langchain-ollama>=0.2.0
langchain-openai>=0.3.1
langchain-text-splitters>=0.3.0
pillow

# Utils
requests
tqdm
